discount scheduler functionality
