+++ app/components/DiscountScheduleEditor/messages.js
@@ -1,0 +1,13 @@
+/*
+ * DiscountScheduleEditor Messages
+ *
+ * This contains all the text for the DiscountScheduleEditor component.
+ */
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  header: {
+    id: 'app.components.DiscountScheduleEditor.header',
+    defaultMessage: 'This is the DiscountScheduleEditor component !',
+  },
+});
