+++ app/components/DiscountScheduleEditor/tests/index.test.js
@@ -1,0 +1,10 @@
+// import React from 'react';
+// import { shallow } from 'enzyme';
+
+// import DiscountScheduleEditor from '../index';
+
+describe('<DiscountScheduleEditor />', () => {
+  it('Expect to have unit tests specified', () => {
+    expect(true).toEqual(false);
+  });
+});
